requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
joblib==1.4.2
pyyaml==6.0.2
prometheus-client==0.20.0
python-json-logger==2.0.7
requests==2.32.3

# Monitoring / Drift (optional but enabled in docker-compose monitoring image)
evidently==0.4.33
